requirements.txt
affine==2.3.0
attrs==19.1.0
boto3==1.10.44
botocore==1.13.44
branca==0.3.1
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
cycler==0.10.0
decorator==4.4.0
dill==0.3.0
Django==1.11.25
django-bootstrap==0.2.4
django-bootstrap4==0.0.8
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-fontawesome==1.0
django-leaflet==0.24.0
docutils==0.15.2
enum34==1.1.6
et-xmlfile==1.0.1
Fiona==1.8.6
folium==0.8.3
functools32==3.2.3.post2
future==0.15.2
futures==3.3.0
GDAL==2.2.4
geocoder==1.38.1
geojson==2.4.1
jdcal==1.4.1
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
matplotlib==1.5.2
mpmath==0.19
multiprocess==0.70.8
munch==2.3.2
nose==1.3.7
numpy==1.9.3
openpyxl==2.6.2
pandas==0.18.1
pathos==0.2.4
Pillow==6.2.1
pox==0.2.6
ppft==1.6.6.1
psycopg2==2.8.3
pyodbc==4.0.26
pyparsing==2.4.5
pypiwin32==223
pyproj==1.9.6
python-dateutil==2.5.3
pytz==2016.6.1
pywin32==224
PyYAML==5.1.1
rasterio==1.1.1
ratelim==0.1.6
requests==2.11.1
s3transfer==0.2.1
scipy==0.17.0
Shapely==1.6.4.post2
six==1.10.0
snuggs==1.4.7
sympy==1.0
urllib2
virtualenv==16.7.0
xlrd==1.0.0
XlsxWriter==1.1.8
xlwt==1.1.2
